# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/project/queryTask": {
            "post": {
                "description": "Post 请求, 发送表单数据, 参数在消息体中",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projectApi"
                ],
                "summary": "Post 请求, 发送表单数据, 参数在消息体中",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "项目 名称",
                        "name": "project_name",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "项目发布者id",
                        "name": "project_publisher_id",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "项目预算最小值",
                        "name": "min_project_budget",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "项目预算最大值",
                        "name": "max_project_budget",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "项目周期",
                        "name": "project_duration",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "创建时间开始",
                        "name": "create_time_begin",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "创建事件结束",
                        "name": "create_time_end",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "最后一次更新时间开始",
                        "name": "update_time_begin",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "最后一次更新时间结束",
                        "name": "update_time_end",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "项目进度ID",
                        "name": "project_schedule_id",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "工作类型",
                        "name": "work_type",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "description": "项目所需工程师类型列表",
                        "name": "project_position_type_list",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "description": "项目类型列表",
                        "name": "project_type_list",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "排序类型 0为按金额排序 1为按周期排序",
                        "name": "order_type",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "排序方式 0为正序 1为逆序",
                        "name": "order_way",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/project/sendTask": {
            "post": {
                "description": "Post 请求, 发送 Json 数据, 参数在消息体中",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projectApi"
                ],
                "summary": "Post 请求, 发送 Json 数据, 参数在消息体中",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目名称",
                        "name": "project_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "项目预算",
                        "name": "project_budget",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "项目简介",
                        "name": "project_short_intro",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "项目logo",
                        "name": "ProjectLogo",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "项目介绍",
                        "name": "ProjectIntroduction",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "项目周期",
                        "name": "ProjectDuration",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "项目开始时间",
                        "name": "StartTime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "项目结束时间",
                        "name": "EndTime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "工作内容说明",
                        "name": "WorkExplain",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "description": "工作类型说明",
                        "name": "ProjectTypeIdList",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "description": "工程师类型ID列表",
                        "name": "EngineerTypeIdList",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "工作类型",
                        "name": "WorkType",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "项目文件",
                        "name": "FileList",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "swagger.Http400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "msg": {
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
        "swagger.Http404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "msg": {
                    "type": "string",
                    "example": "Page Not Found"
                }
            }
        },
        "swagger.Http500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "msg": {
                    "type": "string",
                    "example": "服务器内部错误"
                }
            }
        },
        "swagger.HttpOk": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "msg": {
                    "type": "string",
                    "example": "接口调用成功"
                }
            }
        }
    }
}